feat: layout에 따라 컴포넌트 분리
